refactor: 💡 Working on Modularization
