+++ .vscode/settings.json
@@ -1,3 +1,4 @@
 {
-    "python.pythonPath": "env\\Scripts\\python.exe"
+    "python.pythonPath": "env\\Scripts\\python.exe",
+    "python.formatting.provider": "black"
 }
