+++ plugins/TodayAtMun/__init__.py
@@ -1,13 +1,9 @@
 import discord
 from discord.ext import commands
-from plugins.TodayAtMun.Month import month
-from datetime import datetime
-from datetime import timedelta
-from json import load
 from Plugin import AutomataPlugin
 from Bot import bot
 from plugins.TodayAtMun.diary_parser import diary_parser
-from pathlib import Path
+from plugins.TodayAtMun.today import Today
 
 
 class TodayAtMun(AutomataPlugin):
@@ -18,40 +14,8 @@ class TodayAtMun(AutomataPlugin):
     def __init__(self, manifest, bot):
         super().__init__(manifest, bot)
         self.result = ""
-        path = Path(__file__).parent
-        file_name = path / "diary.json"
-        with open(file_name, "r") as f:
-            self.info = load(f)
-        self.set_current_date()
-
-    def set_current_date(self):
-        """ Current Day, Month, Hour, second """
-        self.date = datetime.now()
-
-    def format_date(self, date):
-        """Provides Current Date formatted to Muns style."""
-        temp = date.strftime("%Y-%#m-%#d-%A").split("-")
-        currYear = temp[0]
-        currMonth = int(temp[1])
-        currDay = temp[2]
-        currDayOfWeek = temp[3]
-
-        return f"{month[currMonth]} {currDay}, {currYear}, {currDayOfWeek}"
-
-    def nextDay(self):
-        """Increases Day By One """
-        self.date = self.date + timedelta(days=1)
-        return self.date
-
-    def findEvent(self, date):
-        """ Provides the significant event on the mun calendar """
-        self.fdate = self.format_date(self.date)
-        for key in self.info:
-            if key == self.fdate:
-                self.infoDay = self.info[key]
-                return
-
-        self.findEvent(self.nextDay())
+        self.ins_Today = Today()
+        self.ins_Today.set_current_date()
 
     @commands.command()
     async def today(self, ctx, arg=None):
@@ -68,11 +32,11 @@ class TodayAtMun(AutomataPlugin):
             await ctx.send("Data Reset.")
             return
 
-        self.set_current_date()
-        self.findEvent(self.date)
+        self.ins_Today.set_current_date()
+        self.ins_Today.findEvent(self.ins_Today.date)
         embed = discord.Embed(
-            title=f"{self.format_date(self.date)}",
-            description=f"```{self.infoDay}```",
+            title=f"{self.ins_Today.format_date(self.ins_Today.date)}",
+            description=f"```{self.ins_Today.infoDay}```",
             url="https://www.mun.ca/regoff/calendar/sectionNo=GENINFO-0086",
             colour=discord.Colour.orange(),
         )
@@ -83,4 +47,10 @@ class TodayAtMun(AutomataPlugin):
         """
         Sends (n) amounts of dates coming up on the univsersity diary
         """
-        pass
+        self.ins_Today.set_current_date()
+        self.ins_Today.findEvent(self.date)
+        self.ins_Today.goToEvent()
+        self.set = iter(self.info)
+        for i in range(arg):
+            setOfDates = next(self.set)
+            await ctx.send(setOfDates)
